automation/requirements.txt
flask==2.3.3
flask-cors==4.0.0
celery==5.3.4
redis==5.0.1
ansible==8.5.0
ansible-core==2.15.5
paramiko==3.3.1
netmiko==4.2.0
requests==2.31.0
pyyaml==6.0.1
jinja2==3.1.2
cryptography==41.0.7
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
